requirements.txt
# Stream Deck library
streamdeck>=0.9.0

# Image processing
Pillow>=10.0.0

# Filesystem monitoring
watchdog>=3.0.0